api/app.py
from flask import Flask, request, jsonify
from decouple import config
import os

app = Flask(__name__)

# Configuração
GEMINI_API_KEY = config('GEMINI_API_KEY')

# Rotas para cada estratégia
@app.route('/api/tool-augmented', methods=['POST'])
def tool_augmented():
    data = request.json
    # Implementação futura
    return jsonify({"message": "Tool-Augmented Reasoning endpoint"})

@app.route('/api/chain-of-thought', methods=['POST'])
def chain_of_thought():
    data = request.json
    # Implementação futura
    return jsonify({"message": "Chain-of-Thought endpoint"})

@app.route('/api/rationale', methods=['POST'])
def rationale():
    data = request.json
    # Implementação futura
    return jsonify({"message": "Rationale Engineering endpoint"})

@app.route('/api/in-context', methods=['POST'])
def in_context():
    data = request.json
    # Implementação futura
    return jsonify({"message": "In-Context Learning endpoint"})

@app.route('/api/finetuning', methods=['POST'])
def finetuning():
    data = request.json
    # Implementação futura
    return jsonify({"message": "Finetuning Simulado endpoint"})

@app.route('/api/memory', methods=['POST'])
def memory():
    data = request.json
    # Implementação futura
    return jsonify({"message": "Memória e Contexto endpoint"})

@app.route('/api/protocol', methods=['POST'])
def protocol():
    data = request.json
    # Implementação futura
    return jsonify({"message": "Protocolo de Contexto endpoint"})

@app.route('/api/all', methods=['POST'])
def all_strategies():
    data = request.json
    # Implementação futura
    return jsonify({"message": "Todas as estratégias combinadas"})

if __name__ == '__main__':
    app.run(debug=True) 