requirements.txt
flask==3.0.0
streamlit==1.32.0
python-dotenv==1.0.0
requests==2.31.0
google-cloud-aiplatform==1.38.1
python-decouple==3.8 